Split existing tests
