+++ site/gatsby-site/cypress/e2e/integration/apps/reports.cy.js
