+++ site/gatsby-site/cypress/e2e/integration/discover.cy.js
@@ -1,5 +1,5 @@
-import flaggedReport from '../fixtures/reports/flagged.json';
-import unflaggedReport from '../fixtures/reports/unflagged.json';
+import flaggedReport from '../../fixtures/reports/flagged.json';
+import unflaggedReport from '../../fixtures/reports/unflagged.json';
 
 describe('The Discover app', () => {
   const url = '/apps/discover';
