+++ site/gatsby-site/cypress/e2e/integration/entity.cy.js
@@ -1,7 +1,7 @@
-import { maybeIt } from '../support/utils';
-import emptySubscriptionsData from '../fixtures/subscriptions/empty-subscriptions.json';
-import subscriptionsData from '../fixtures/subscriptions/subscriptions.json';
-const { SUBSCRIPTION_TYPE } = require('../../src/utils/subscriptions');
+import { maybeIt } from '../../support/utils';
+import emptySubscriptionsData from '../../fixtures/subscriptions/empty-subscriptions.json';
+import subscriptionsData from '../../fixtures/subscriptions/subscriptions.json';
+const { SUBSCRIPTION_TYPE } = require('../../../src/utils/subscriptions');
 
 const entity = {
   entity_id: 'google',
