+++ site/gatsby-site/cypress/e2e/integration/landingPage.cy.js
@@ -1,4 +1,4 @@
-import { maybeIt } from '../support/utils';
+import { maybeIt } from '../../support/utils';
 const { format } = require('date-fns');
 
 describe('The Landing page', () => {
