+++ site/gatsby-site/cypress/e2e/integration/translationBadge.cy.js
