+++ site/gatsby-site/cypress/e2e/unit/Translator.cy.js
@@ -1,6 +1,6 @@
 const { ObjectID } = require('bson');
 
-const Translator = require('../../src/utils/Translator');
+const Translator = require('../../../src/utils/Translator');
 
 const reports = [
   {
