+++ site/gatsby-site/cypress/e2e/unit/functions/onNewIncident.cy.js
@@ -1,6 +1,6 @@
-const { SUBSCRIPTION_TYPE } = require('../../../src/utils/subscriptions');
+const { SUBSCRIPTION_TYPE } = require('../../../../src/utils/subscriptions');
 
-const onNewIncident = require('../../../../realm/functions/onNewIncident');
+const onNewIncident = require('../../../../../realm/functions/onNewIncident');
 
 const subscriptionsToNewIncidents = [
   {
